[domain] create domain package & move entity
