+++ src/main/java/com/humming/domain/ConcertStatus.java
@@ -1,4 +1,4 @@
-package com.humming.core.concert;
+package com.humming.domain;
 
 import lombok.Getter;
 import lombok.RequiredArgsConstructor;
